Refactor event statistics (analytics) to support `react-router` v6
