+++ app/routes/events/components/Analytics.css
@@ -1,0 +1,22 @@
+.metrics {
+  height: 80px;
+  line-height: 1.3;
+  margin-bottom: 1rem;
+}
+
+.metricHeader {
+  font-weight: 500;
+  text-transform: uppercase;
+  color: var(--secondary-font-color);
+}
+
+.metricNumber {
+  font-size: var(--font-size-xl);
+  font-weight: 500;
+}
+
+.metricDivider {
+  width: 2px;
+  height: 100%;
+  background-color: var(--border-gray);
+}
